feat: update i18n integration and refactor error handling with Zod
