+++ packages/shared/src/models/generalPreference.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from '@server/i18n';
 
 export enum Appearance {
   LIGHT = 'light',
