+++ packages/shared/src/models/user.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from '@server/i18n';
 
 export enum Role {
   USER = 'user',
